Add block reinforcement system with event listeners
Introduces a block reinforcement system allowing blocks to be lightly or heavily reinforced, tracked per-chunk using a new Reinforcement record. Adds listeners for right-click, burn, and explosion events to manage reinforcement interactions. Updates block breaking logic to account for reinforcement multipliers and ensures reinforcements are removed when blocks are broken. Refactors PlaceBlockListener to use the new system and registers all new listeners in the plugin main class. Also adds configuration options for reinforcement multipliers.

+++ src/main/java/com/minecraftcivilizations/specialization/GUI/ClassGUI.java
@@ -187,7 +187,7 @@ public class ClassGUI extends GUI {
             }
 
             double currentXp = Math.round(skill.getXp() * 100) / 100D;
-            double percentOfTotalForNextLevel = Math.round( (Double) SpecializationConfig.getSkillRequirementsConfig().get(skill.getSkillType() + "_" + SkillLevel.getSkillLevelFromInt(currentSkillLevel + 1) + "_REQUIREMENT", Double.class) * 100) / 100D;
+            double percentOfTotalForNextLevel = Math.round(SpecializationConfig.getSkillRequirementsConfig().get(skill.getSkillType() + "_" + SkillLevel.getSkillLevelFromInt(currentSkillLevel + 1) + "_REQUIREMENT", Double.class) * 100) / 100D;
             double xpToNextLevel = Math.max(
                     Math.round((Skill.getXPNeededForLevel(currentSkillLevel + 1) - skill.getXp()) * 100) / 100D,
                     Math.round((percentOfTotalForNextLevel / 100 * customPlayer.getTotalXp() - currentXp) / (1.0 - percentOfTotalForNextLevel / 100))
